authentication.r file
An R Script "authentication.r" written to authorize R to use twitter API.  The R script is run in the R console-asking to enter the pin to authorize R. The R console opens a web page in a browser. After clicking on the authorize app button, the pin is displayed in the browser. Copy the pin to the R console at the cursor. Saving the authorization credential object into a .Rdata file.